Module binjs_generic::syntax

Walkers, comparisons.

Structs

ASTWalker

Comparator

MutASTWalker

Enums

ASTError

Traits

ASTVisitor

Compare
Compare two values JSON as elements of a given spec. Fields that are not part of the spec are ignored.

MutASTVisitor

Functions

compare
Compare two ASTs, restricting comparison to the items that appear in the grammar.

validate
Ensure that a value is an inhabitant of the grammar.

Type Definitions

WalkPath

WalkPathItem